Optimize React code stage-1
